requirements.txt
# Core dependencies
pyshark==0.6
scapy==2.5.0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.13.0
torch==2.1.0
torchvision==0.16.0

# Data processing
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# CLI
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
joblib==1.3.2
python-dateutil==2.8.2
requests==2.31.0

# Logging
loguru==0.7.2

# Security
cryptography==41.0.7
